requirements.txt
Django==2.0.4
django-debug-toolbar==1.9.1
django-model-utils==3.1.1
djangorestframework==3.8.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.1
pytz==2018.3
sqlparse==0.2.4
